+++ web/package.json
@@ -12,14 +12,12 @@
     "flag-icon-css": "^3.2.1",
     "formik": "^1.5.1",
     "humps": "^2.0.1",
-    "i18next": "^15.0.6",
     "jwt-decode": "^2.2.0",
     "lodash": "^4.17.11",
     "prop-types": "^15.7.2",
     "ramda": "^0.26.1",
     "react": "^16.8.3",
     "react-dom": "^16.8.3",
-    "react-i18next": "^10.3.0",
     "react-intl": "^2.8.0",
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.5",
