+++ web/src/components/atomics/index.js
@@ -1,3 +1,2 @@
 export { Text } from './text'
 export { Box } from './box'
-export { LinkButton } from './link-button'
