--- web/src/components/layout/navbar/index.js
@@ -1,2 +1,0 @@
-export { Navbar } from './navbar'
-export { UnauthorizedNavbar } from './unauthorized-navbar'
