+++ web/src/screens/layout/navbar/index.js
@@ -1,0 +1,2 @@
+export { default as Navbar } from './navbar'
+export { default as UnauthorizedNavbar } from './unauthorized-navbar'
