+++ web/src/screens/layout/navbar/navbar.module.scss
