+++ web/src/screens/layout/navbar/sign-in-menu/sign-in-menu.module.scss
