+++ web/src/screens/layout/navbar/user-menu.module.scss
